Moved statefile reader to utilities
